requirements.txt
gradio==4.44.0
langchain>=0.2.17
langchain-google-genai>=2.1.8
langchain-core>=0.2.43
python-dotenv==1.0.0 